backend/pyproject.toml
[tool.poetry]
name = "ispas-backend"
version = "0.1.0"
description = "Intelligent Stock Price Analysis System - Backend API"
authors = ["ISPAS Team"]
readme = "README.md"
python = "^3.14"

[tool.poetry.dependencies]
python = "^3.14"
fastapi = "^0.129.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
pydantic = "^2.10.0"
pydantic-settings = "^2.7.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
alembic = "^1.14.0"
asyncpg = "^0.30.0"
redis = {extras = ["hiredis"], version = "^5.2.0"}
aiohttp = "^3.11.0"
pandas = "^2.2.0"
numpy = "^2.2.0"
python-dotenv = "^1.0.1"
tenacity = "^9.0.0"
celery = {extras = ["redis"], version = "^5.4.0"}

# 머신러닝
torch = "^2.10.0"
tensorflow = "^2.20.0"
scikit-learn = "^1.6.0"
xgboost = "^2.1.0"
optuna = "^4.1.0"
shap = "^0.46.0"

# 기술적 지표
ta-lib-bin = {version = "^0.4.28", markers = "sys_platform == 'win32'"}
pandas-ta = "^0.3.14"

# 데이터 수집
yfinance = "^0.2.50"

# 유틸리티
structlog = "^24.4.0"
python-multipart = "^0.0.20"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
ruff = "^0.8.0"
black = "^24.10.0"
mypy = "^1.14.0"
ipython = "^8.31.0"
jupyter = "^1.1.0"

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py314"]

[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
